Move 2 reference-related typeclasses to label 'type's
