+++ code/drasil-lang/lib/Language/Drasil/DecoratedReference.hs
@@ -11,7 +11,7 @@ module Language.Drasil.DecoratedReference (
 
 import Language.Drasil.Sentence (RefInfo(..))
 import Language.Drasil.Reference (Reference, ref)
-import Language.Drasil.Classes.Core (HasRefAddress(..))
+import Language.Drasil.Label.Type (HasRefAddress(..))
 import Language.Drasil.ShortName (HasShortName(..))
 import Language.Drasil.UID (HasUID(..))
 import Control.Lens ((^.), makeLenses, Lens')
