+++ code/drasil-lang/lib/Language/Drasil/Label/Type.hs
@@ -3,13 +3,15 @@
 module Language.Drasil.Label.Type(
   -- * Types
     LblType(RP, Citation, URI), IRefProg(..)
+  -- * Classes
+  , HasRefAddress(..), Referable(..)
   -- * 'LblType' accessor
   , getAdd
   -- * 'IRefProg' constructors
   , name, (+::+), raw, defer, prepend
-  ) where
+) where
 
-import Language.Drasil.UID (UID)
+import Language.Drasil.UID (UID, HasUID)
 
 -- | Applying different pieces of information for a reference.
 -- An RP is a decorated internal reference.
@@ -26,6 +28,19 @@ data IRefProg =
   | RS String                   -- ^ Lifts a 'String' into a 'RefProg'.
   | RConcat IRefProg IRefProg   -- ^ Concatenates with two subprograms.
   | Name                        -- ^ The 'Symbol' to insert the 'ShortName' directly.
+
+-- | Members must have a reference address.
+class HasRefAddress b where
+  -- | Provides the ability to hold a reference address.
+  getRefAdd :: b -> LblType
+
+-- | Members of this class have the ability to be referenced.
+class (HasUID s, HasRefAddress s) => Referable s where
+  -- | The referencing address (what we're linking to).
+  -- Only visible in the source (tex/html).
+  refAdd    :: s -> String 
+  -- | Alternate form of reference.
+  renderRef :: s -> LblType 
 
 -- | Retrieves the 'String' contained in a 'LblType'.
 getAdd :: LblType -> String
